refactor(activity_feed): break plugin <-> feed_panel circular import (refs #1367)
Extract build_projector + _collect_work_db_paths into a new leaf module
pollypm.plugins_builtin.activity_feed.projector_factory. Both plugin.py
and cockpit/feed_panel.py now import from the leaf; the lazy import of
new_event_count inside the badge provider is gone. plugin.py re-exports
build_projector so external callers (cockpit_ui, cockpit_inbox, cli)
are unaffected.

One pair down. Remaining cycle pairs from #1367: cockpit_inbox/
cockpit_inbox_items, cockpit_palette/cockpit_ui, cockpit_rail/cockpit,
onboarding/onboarding_tui, work.service_notifications/sqlite_service,
work.service_queries/sqlite_service.
